+++ src/oclaunch.ml
@@ -46,6 +46,5 @@ let version_number = "0.2.2";;
 let build_info = ( "Build with OCaml version " ^ (Sys.ocaml_version) ^ " on " ^ (Sys.os_type) );;
 
 let () =
-  Messages.debug "Started";
   Command.run ~version:version_number ~build_info:build_info Command_def.commands
 ;;